Fixed bootloader
Fixed errors in the bootloader. I'm embarrassed to say I only just discovered how to properly debug with bochs and immediately saw the problem. In short, the code generated for the helper routines was 32-bits because that was the last directive nasm encountered. Now the kernel is loaded into memory and reached by the bootloader. Next is to properly link everything to get rid of all of these magic numbers scattered around.