Reverted
Found that the 'seperation of concerns' approach requires the data within 'boot.s' to also be split into its own file because the assembler creates absolute offsets for labels, making them all wrong once the binary is loaded into memory. The ORG directive cannot be used either when assembling an ELF file. I decided it was better to have one large file instead of yet another file for the data.